Reorganise PaletteColorPicker
