+++ content/webapp/components/PaletteColorPicker/PaletteColorPicker.HueSlider.tsx
